Bump to go1.16
