+++ .travis.yml
@@ -1,8 +1,8 @@
 language: go
 
 go:
-  - 1.14.x
   - 1.15.x
+  - 1.x
 
 go_import_path: github.com/vulcand/oxy
